+++ src/control_plane/migration/volume/mod.rs
@@ -1,8 +1,12 @@
 mod backup_project_volume;
 mod project_volume_backup_options;
+mod project_volume_restore_options;
+mod restore_project_volume;
 
 #[cfg(test)]
 mod tests;
 
 pub(crate) use backup_project_volume::backup_project_volume;
 pub(crate) use project_volume_backup_options::ProjectVolumeBackupOptions;
+pub(crate) use project_volume_restore_options::ProjectVolumeRestoreOptions;
+pub(crate) use restore_project_volume::restore_project_volume;